Move Cloud-specific methods into new command manager class
